docextract/backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
python-dotenv
paddlepaddle==2.6.2
paddleocr==2.7.3
PyMuPDF
Pillow
llama-parse
google-genai
python-docx
numpy==1.26.4